tests: fold the actionAngleTorusStaeckel tests into test_actionAngle.py
The CI matrix enumerates test files explicitly and cannot be extended
from this token (workflow scope), so the torus-mapper tests move into
the file the matrix already runs -- where every other actionAngle test
lives anyway. This is what codecov/patch was failing on: the new module
was tested locally but its test file never ran in CI.
